Commands are set up to send to another device using BT
The SettingsActivity currently contains a button that sends a command,
"FMCW:Bandwidth ? CR", to another Android device that has established a
Bluetooth connection prior. This Android device should be currently
running the BluetoothChat application, and the UUID must contain the
MAC address for a Secure Connection. However, I'm assuming that as long
as both applications on each device are hosting the same UUID, then the
devices/application should be able to setup a wireless connection with
each other.