Communicated with Radar kit for the 1st time!
A lot of adjustments needed...

#1 important change to remember:
--> changed UUID from INSECURE to SPP !!

Other important changes:
- Parsing Commands (Return methods: String(String str1),
List<String>(String str1) , String(String str1, boolean queryradar)

--> List<String> :
ATTEMPT to read commands sent from Radar Kit to application and puts
strings into a List based on whether they are split by the 'end command
character'
--> 'end command character' :
used to be \n then was \r\n and now currently a semicolon (based on
zach's decison"
** IMPORTANT DISCOVERED KNOWLEDGE **
:: Using the Debug mode :: Noticed that each character was saved into
an entry in a byte array (byte[1024]) :: ASCII values were used to
reference characters :: NOTED that Java already interprets the ASCII
characters for Carriage Return (ascii = 13) and New Line(ascii = 10)
for each command sent from Radar kit to the Android Application.

HENCE, Adam does not need to send back these character values in the
strings sent from the Kit to the Android device.

--> PIC was [temporarily] programmed to send commands* in a continuous
loop including the phrases "AppliedRadar", "water bottle",
"CarriageReturn\r", and "NewLine\n".
*not actual commands

--> LogCat window interpretted this messages in the beginning of the
loop and then began distorting the commands into separate lines in the
LogCat window, with sometimes half a command, a signle letter, or more
than one (not complete) command in a single line. All command letters
remained in the order they were sent.

2 possible causes:
--> 1. The application (and BluetoothChatService) are [both] not
reading or handling the messages properly
OR
--> 2. There is something going wrong when the PIC is sending commands
to the Android device running my application.

However-- I believe that #1 is more likely to be the direct cause of
the distortion, but the PIC may also be playing a part in the way the
messages are sent.

::: LASTLY :::
When the application sends commands to the Kit, the Kit does NOT at all
respond to the device even though the String commands are as they
should be according to the Commands List Adam gave me [as far as Adam &
I can see]