Added buffer for data sent to application
Assuming the data sent from the kit is sent in a comma-separated list,
the application tests whether there is a comma in the returned string.
If there is NOT any commas, then the 'readCommand' method in the
RadarCommand class is called (which only returns the string back
without any changes to it). If there is AT LEAST ONE comma in the
returned string, then the 'parceCommand' (in the RadarCommand class) is
called to save each element in the list as an element in an array of
strings (without ANY commas).

Also created an unused method that is SUPPOSED to parce the strings in
List<string> into an ArrayList of integers (ArrayList<Integer>).