Tablet-to-Tablet Communication to Simulate Data Transfer
UUIDs were changed to support the Tablet communication for both the
FMCW Radar App (Tablet2), as well as the "BluetoothChat_Evan"
application (Tablet1).

Tab1 can send comma-separated values (CSVs) to Tab2, which will read in
the "messages" (CSVs) that are type String. From this point, the CSVs
are separated into an array of strings and from there converted into an
array of integers. Once convert, for testing purposes, the indices and
their values are printed to the LogCat to display.