File Archive Beta = Success!;  Storing & Loading "Data" received from the tablet1 (BTChat App) into App = Success!
File Archive is set up to display "previously collected data" that was
stored upon recieving the CSVs from Tablet1 (using the BTChat App). Any
file within the "/mnt/sdcard/FMCW File Archive" directory is displayed
here. When a file is selected (onItemClickListener), it is loaded into
an array of [potentially] plottable data.

** Currently only saves and plots integer values that are recieved in
CSV-form (from Tablet1, as of now) **

Once Tablet2 recieves this data, it is automatically being saved into a
file within the "/mnt/sdcard/FMCW File Archive" directory.
--> It's not very logical, but it works for now.

This recieved data is ONLY ploted when the user goes into the File
Archive and selects a file to plot.
--> Once again, not very logical, but I was tired last night and wasn't
thinking logically enough to implement how it should [logically] work.

Lots of sloppy and useless code from last semester, regarding MAINLY
[but really everything] plotting raw data, plotting the FFT data,
saving the FFT Data [*need to decipher*], and "loading" the data from
file!

So --> yesterday, instead going back and fixing LOTS of ugly and
impractical coding, I just created new methods. These methods worked
with the integer [CSVs} recieved from the BTChat App, but are actually
somewhat PRACTICAL. They are general functions, so they support more
than just one file of data and the length of the data can vary, which
is the most important factor (with regard to the previous, pathetically
working implementation of reading data out of a file into an array).

** Probably already inferred, but -- App WILL CRASH if the "Save"
button is selected (which is fine, app is still okay), however, then if
that file is selected to plot in GUI, app crashes due to FLOATING-POINT
VALUES! **

--> Need to figure out how to implement float types.
--> Need to HARDCORE clean up ugly, pathetic code.
--> Need to be more specific regarding the contents of a file --
whether it is RAW data being saved or FFT data. *Important!*