Added check to decipher between parameters commands returned & data
Used a boolean "commandInfo" to dipher what kind of data is being
returned to the application upon requested.

-> commandInfo is orginally initialized as "false"
-> commandInfo is set to "true" ONLY when a command is sent to the
radar to "get" current radar parameter information
-> boolean value doesn't not change when the "Collect" button/command
is sent to the radar requesting data to be returned to device
-> value also doesn't not change when a command is sent to the radar to
"set" radar parameters (to something other than the default, or
current, settings)

:: BUGS ::
- kit is currently setup (for some unknown reason) to send "END" twice
whenever is returns radar parameter settings to [Android] device
- the "END" is supposed to only send once, just like the kit does when
it sends data to the application
--> our application only checks for the one, or the first, "END\n\r"
detection bits, therefore, if parameter settings are requested and
returned, our application only looks for first end-bits, and then sends
to the MainActivity, leaving the second "END\n\r" bits in the byte
buffer in the BluetoothChatService class.
* This causes application to "process" these 5 bits that aren't
relevant to any of our information when we read from the byte buffer
again.