User input setup; Parameter commands updated; Evan's defaults para, working?
DialogFragment : implemented for user to input specific parameter
settings for radar kit.
- So far, only set up for user to enter the ramptime

Currently:
- user clicks 'set ramptime' button
- DialogFragment pops up
- User enters an integer value -> we hope ...
- User has options to 'Cancel' (no input) or 'OK' the input
- Clicking 'OK' button in DiaglogFrag displays the text entered in DF
in a TextView in the SettingsActivity
- User can select 'Send' button to send the command to the radar kit
over Bluetooth

Radar Parameter Commands : updated
- ONLY* needs dollar sign, '$', to detect that a command has been sent
to the kit; Nothing more, nothing less (than the $)
- Had issues because there were '$\n' at the end of the commands in the
code. --> REMOVED ALL OF THEM

Evan's code for transmitting 'getter' commands prompty as devices are
connected via BT :: added to this code
- Not sure what the status of this is
- Not sure if we are receiving anything and "attempting" to handle it
(or if we are and its just not be returned to SettingsActivity)
- Almost positive this is only occuring onClick of the 'Collect/Save'
button in MainActivity.
--> Not actually transmitting commands on start of the device Bluetooth
connection.