GUI Appearance updates - file archive mainly.
Mostly appearance updates of the application as a whole.

Changed the file archive a lot, allows for more than one document to be
saved during the time of the last document saved.
--> remove code regarding file kind/types.

Temporarily commented the portion of code that saves the FFT processed
data since I dont want any accidental issues in the presentation demo
on Monday.

Added comments/removed unused code. Still lots of unused code through
out the application, but for some reason a warning isn't appearing for
that stuff.